backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
PyPDF2
python-multipart
python-dotenv